+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Next Release
 * Your contribution here.
+* Add `Miron::Middleware::SSL` to force requests to SSL - [@maclover7](https://github.com/maclover7).
+* Add `HTTPS` hash key to Miron Specification - [@maclover7](https://github.com/maclover7).
+* Add `HTTP_HOST` hash key to Miron Specification - [@maclover7](https://github.com/maclover7).
 
 ### 0.0.4 - (10/17/2015)
 * Refactor request processing to use `Miron::RequestFetcher`, and not